+++ plugins/O-Formant/Source/PluginEditor.h
@@ -41,6 +41,7 @@ private:
     std::unique_ptr<juce::WebSliderRelay> formantShiftRelay;
     std::unique_ptr<juce::WebSliderRelay> formantSpreadRelay;
     std::unique_ptr<juce::WebSliderRelay> pitchGlideRelay;
+    std::unique_ptr<juce::WebSliderRelay> transitionTimeRelay;
     std::unique_ptr<juce::WebSliderRelay> outputGainRelay;
     std::unique_ptr<juce::WebSliderRelay> stereoWidthRelay;
 
@@ -72,6 +73,7 @@ private:
     std::unique_ptr<juce::WebSliderParameterAttachment> formantShiftAttachment;
     std::unique_ptr<juce::WebSliderParameterAttachment> formantSpreadAttachment;
     std::unique_ptr<juce::WebSliderParameterAttachment> pitchGlideAttachment;
+    std::unique_ptr<juce::WebSliderParameterAttachment> transitionTimeAttachment;
     std::unique_ptr<juce::WebSliderParameterAttachment> outputGainAttachment;
     std::unique_ptr<juce::WebSliderParameterAttachment> stereoWidthAttachment;